backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Async & WebSocket
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Task Queue
celery>=5.3,<6.0
redis>=5.0,<6.0

# Authentication
django-allauth>=0.61,<1.0
djangorestframework-simplejwt>=5.3,<6.0

# Two-Factor Authentication
pyotp>=2.9,<3.0
qrcode>=7.4,<8.0

# Storage
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Production Server
gunicorn>=21.0,<23.0

# Utilities
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
requests>=2.31,<3.0

# mkv2cast video converter
mkv2cast>=1.1.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-asyncio>=0.21,<1.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
